refactor: background
